INC-224 refactor: extract UserSearchItems into standalone component
