+++ src/components/pagesComponents/userSearch/UserSearchPage.tsx
@@ -1,53 +1,23 @@
-import { ChangeEvent, useEffect, useRef, useState } from 'react'
+import { ChangeEvent, useState } from 'react'
 
-import { useDebounce } from '@/common/hooks/useDebounce'
-import { useElementInView } from '@/common/hooks/useElementInView'
 import { useTranslation } from '@/common/hooks/useTranslation'
-import { useGetUsersProfilesQuery } from '@/services/usersFollowingAndFollowersService'
 import { TextField, Typography } from '@samuraichikit/inc-ui-kit'
 
 import s from './userSearchPage.module.scss'
 
-import { UserSearchItem } from './userSearchItem'
+import { UserSearchItems } from './userSearchItems'
 
 export const UserSearchPage = () => {
   const classNames = {
     textField: s.textField,
-    usersProfilesContainer: s.usersProfilesContainer,
   }
+
   const [search, setSearch] = useState('')
-  const [cursor, setCursor] = useState(0)
   const { t } = useTranslation()
-  const { isInView, targetRef } = useElementInView({ threshold: 0.5 })
-  const prevInViewRef = useRef(false)
-  const debouncedSearch = useDebounce(search)
-
-  const { data } = useGetUsersProfilesQuery(
-    { cursor, pageSize: 14, search: debouncedSearch },
-
-    { skip: !debouncedSearch }
-  )
-  const usersProfiles = data?.items
-  const nextCursor = data?.nextCursor
-  const itemsToRender = debouncedSearch ? usersProfiles : []
 
   const searchHandler = (e: ChangeEvent<HTMLInputElement>) => {
     setSearch(e.currentTarget.value)
   }
-
-  useEffect(() => {
-    prevInViewRef.current = false
-    setCursor(0)
-  }, [debouncedSearch])
-
-  useEffect(() => {
-    const becameVisible = isInView && !prevInViewRef.current
-
-    if (becameVisible && nextCursor) {
-      setCursor(nextCursor)
-    }
-    prevInViewRef.current = isInView
-  }, [isInView, nextCursor])
 
   return (
     <>
@@ -61,19 +31,7 @@ export const UserSearchPage = () => {
         type={'search'}
         value={search}
       />
-      <div className={classNames.usersProfilesContainer}>
-        {itemsToRender?.map(({ avatars, firstName, id, lastName, userName }, index) => (
-          <UserSearchItem
-            firsName={firstName}
-            id={id}
-            key={id}
-            lastName={lastName}
-            ref={index === itemsToRender.length - 1 ? targetRef : null}
-            src={avatars[0]?.url}
-            userName={userName}
-          />
-        ))}
-      </div>
+      <UserSearchItems search={search} />
     </>
   )
 }
