+++ src/components/pagesComponents/userSearch/userSearchItems/index.ts
@@ -1,0 +1,1 @@
+export * from './UserSearchItems'
