+++ src/components/pagesComponents/userSearch/userSearchItems/userSearchItems.module.scss
@@ -1,0 +1,5 @@
+.usersProfilesContainer {
+  display: flex;
+  flex-direction: column;
+  gap: 12px;
+}
