+++ src/components/pagesComponents/userSearch/userSearchPage.module.scss
@@ -1,9 +1,3 @@
 .textField {
   margin: 13px 0 25px;
 }
-
-.usersProfilesContainer {
-  display: flex;
-  flex-direction: column;
-  gap: 12px;
-}
